Develop

+++ src/main/java/com/teddybear6/toegeungil/hobby/controller/HobbyController.java
@@ -412,40 +412,58 @@ public class HobbyController {
 
 
     //후기 삭제
+    @PreAuthorize("hasAnyRole('USER','ADMIN','TUTOR')")
     @DeleteMapping("/review/{reviewCode}")
     @ApiOperation(value = "취미 후기 삭제 Api", notes = "취미 리뷰 번호로 해당 게시글의 참여 후기를 삭제한다.")
-    public ResponseEntity<?> removeReview(@PathVariable int reviewCode) {
+    public ResponseEntity<?> removeReview(@PathVariable int reviewCode, @AuthenticationPrincipal AuthUserDetail userDetails) {
+        Map<String, String> respose = new HashMap<>();
+        int result = 0;
         HobbyReview hobbyReview = hobbyService.findByReviewCode(reviewCode);
         if (Objects.isNull(hobbyReview)) {
             return ResponseEntity.status(404).body("후기가 없습니다.");
         }
-        hobbyReview.setReviewStatus("N");
-        int result = hobbyService.deleteByReviewCode(hobbyReview);
+        if(userDetails.getUserEntity().getUserNo()==hobbyReview.getUserNo()){
+            hobbyReview.setReviewStatus("N");
+            result  = hobbyService.deleteByReviewCode(hobbyReview);
+
+        }
 
         if (result > 0) {
-            return ResponseEntity.ok().body("삭제 완료됐습니다.");
+            respose.put("value","삭제되었습니다.");
+            return ResponseEntity.ok().body(respose);
         } else {
-            return ResponseEntity.status(404).body("삭제 실패했습니다.");
+            respose.put("value","삭제 실패했습니다..");
+            return ResponseEntity.status(404).body(respose);
         }
 
     }
 
     //후기 수정
     @PutMapping("/review/{reviewCode}")
+    @PreAuthorize("hasAnyRole('USER','ADMIN','TUTOR')")
     @ApiOperation(value = "취미 후기 수정 Api", notes = "취미 리뷰 번호로 해당 게시글의 참여 후기를 수정한다.")
-    public ResponseEntity<?> modifyReview(@PathVariable int reviewCode, @RequestBody HobbyReviewDTO hobbyReviewDTO) {
+    public ResponseEntity<?> modifyReview(@PathVariable int reviewCode, @RequestBody HobbyReviewDTO hobbyReviewDTO ,@AuthenticationPrincipal AuthUserDetail userDetails) {
+        Map<String, String> respose = new HashMap<>();
         hobbyReviewDTO.setReviewCode(reviewCode);
         HobbyReview hobbyReview = hobbyService.findByReviewCode(reviewCode);
         if (Objects.isNull(hobbyReview)) {
-            return ResponseEntity.status(404).body("후기가 없습니다.");
+            respose.put("value","후기가 없습니다.");
+            return ResponseEntity.status(404).body(respose);
+        }
+
+        if (hobbyReview.getUserNo()!=userDetails.getUserEntity().getUserNo()) {
+            respose.put("value","작성자가 아닙니다.");
+            return ResponseEntity.status(404).body(respose);
         }
 
         int result = hobbyService.updateReview(hobbyReviewDTO);
 
         if (result > 0) {
-            return ResponseEntity.ok().body("수정 성공했습니다.");
+            respose.put("value","수정 성공했습니다.");
+            return ResponseEntity.ok().body(respose);
         } else {
-            return ResponseEntity.status(500).body("수정 실패했습니다.");
+            respose.put("value","수정 실패했습니다.");
+            return ResponseEntity.status(500).body(respose);
         }
 
 
@@ -544,31 +562,21 @@ public class HobbyController {
     //localhost:8001/hobbys/loacal/1/category/1?page=0&size=5
     @GetMapping("/loacal/{localCode}/category/{categoryCode}")
     @ApiOperation(value = "카테고리 AND 지역 취미 조회 Api", notes = "카테고리 번호와 지역 번호로 두 조건에 모두 해당되는 취미 게시글 목록을 조회한다.")
-    public ResponseEntity<List<?>> localAndCategoryFilter(@PathVariable int localCode, @PathVariable int categoryCode, final Pageable pageable) {
+    public ResponseEntity<Map<String,Object>> localAndCategoryFilter(@PathVariable int localCode, @PathVariable int categoryCode, final Pageable pageable) {
 
 
-        List<HobbyGetDTO> hobbies = hobbyService.findByCategoryCodeAndLocalCode(categoryCode, localCode, pageable);
+        Map<String,Object> hobbies = hobbyService.findByCategoryCodeAndLocalCode(categoryCode, localCode, pageable);
+        List<HobbyGetDTO> hobbyGetDTOS = (List<HobbyGetDTO>) hobbies.get("value");
 
-        if (hobbies.size() == 0) {
-            List<String> error = new ArrayList<>();
-            error.add(null);
+        if (hobbyGetDTOS.size() == 0) {
+            Map<String,Object> error = new HashMap<>();
+            error.put("error",null);
             return ResponseEntity.status(404).body(error);
         }
         return ResponseEntity.ok().body(hobbies);
     }
 
-    @GetMapping("/loacal/size/{localCode}/category/{categoryCode}")
-    @ApiOperation(value = "카테고리 AND 지역 취미 사이즈 조회 Api", notes = "카테고리 번호와 지역 번호로 두 조건에 모두 해당되는 취미 게시들의 사이즈를 조회한다.")
-    public ResponseEntity<?> localAndCategoryFiltersize(@PathVariable int localCode, @PathVariable int categoryCode) {
 
-
-        List<Hobby> hobbies = hobbyService.findByCategoryCodeAndLocalCode(categoryCode, localCode);
-
-        if (hobbies.size() == 0) {
-            return ResponseEntity.status(404).body(0);
-        }
-        return ResponseEntity.ok().body(hobbies.size());
-    }
